Refactor ViewModels to inject polymorphic printer instances for better flexibility
